refactor addressData and move it to utils
